Bug Fix, Account Switching and Depositing now Fully Functional
Received error message; "class bankinggui.TaxFreeSavingsAccount cannot be cast to class bankinggui.RegisteredEducationSavingsPlanAccount" when attempting to use the TFSA feature while it was originally functionally and the casting is intended to be between the TFSA and Debit superclass. The program was simply interpreting this as casting due to a missing break in the switch case.